feat はてなブログ記事Hooks, component作成
