+++ hooks/HatenaArticleList/HetanaArticleListRepository.ts
@@ -5,15 +5,8 @@ import axios from "axios";
 export default class HatenaArticleListRepository {
 
 
-    getArticle = async() => {
-        try{
-            // データタイプを作成して受け取ったデータを使えるようにする
-            const res = await axios.get('api/hatenaArticleList')
-            return res;
-   
-        }catch {
-            console.log("Error");
-        }
+    getArticle = () => {
+        return axios.get('api/hatenaArticleList');
     }
 }
